Use Inheritance instead of Mixins
