--- lib/concerns/base.rb
@@ -1,15 +1,0 @@
-module Concerns
-  module Base
-    def create(name)
-      new(name).save
-    end
-
-    def all
-      class_variable_get :@@all
-    end
-
-    def destroy_all
-      all.clear
-    end
-  end
-end
