--- lib/concerns/saving.rb
@@ -1,8 +1,0 @@
-module Concerns
-  module Saving
-    def save
-      self.class.all << self
-      self
-    end
-  end
-end
